+++ src/components/Main/index.jsx
@@ -1,6 +1,7 @@
 import React, {useState} from 'react';
 import './style.css'; // Importa o CSS
 import Carousel from '../Carousel/Carousel';
+import Header from '../Header/Header'
 
 
 const Main = () => {
@@ -11,31 +12,11 @@ const Main = () => {
         '/images/mias1.jpg',
     ];
 
-    const [isMenuOpen, setIsMenuOpen] = useState(false);
-
-    const toggleMenu = () => {
-        setIsMenuOpen(!isMenuOpen);
-    };
-
+    
     return (
         <div className="App">
-            <header className="header">
-                <img src='/images/logo_sale.png' alt="Logo" className="logo" />
-                <button className="menu-toggle" onClick={toggleMenu}>
-                    &#9776;
-                </button>
-                <nav className={isMenuOpen ? 'nav-open' : ''}>
-                    <ul>
-                        <li className='li-menu'><a href="#ofertas">Ofertas</a></li>
-                        <li className='li-menu'><a href="#precos">Preços</a></li>
-                        <li className='li-menu'><a href="#fan-club">Fan Clube</a></li>
-                        <li className='li-menu'><a href="#contato">Contato</a></li>
-                        <li className='li-menu'><button className="login-btn">Login</button></li>
-                        <li className='li-menu'><button className="register-btn">Registre-se</button></li>
-                    </ul>
-                </nav>
-            </header>
-
+            
+        <Header/>
 
             <main className="main">
                 <section id="ofertas" className="ofertas">
