+++ src/components/Main/style.css
@@ -1,107 +1,3 @@
-/* Cabeçalho (Header) */
-.header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    background-color: #fff;
-    color: #333;
-    padding: 1rem;
-    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-/* @media (max-width: 768px) {
-    .align-header {
-        display: block;
-    }
-} */
-
-.logo {
-    width: 90px;
-}
-
-.menu-toggle {
-    display: none;
-    font-size: 2rem;
-    background: none;
-    border: none;
-    cursor: pointer;
-}
-
-.header nav {
-    flex-grow: 1;
-    display: flex;
-    justify-content: flex-end;
-}
-
-.header nav ul {
-    display: flex;
-    list-style-type: none;
-}
-
-.header nav ul li {
-    margin: 0 1rem;
-}
-
-.header nav ul li a {
-    color: #333;
-    text-decoration: none;
-    font-weight: bold;
-}
-
-.header nav ul li a:hover {
-    color: #FFD700;
-}
-
-.login-btn,
-.register-btn {
-    padding: 0.5rem 1rem;
-    border-radius: 5px;
-}
-
-.login-btn {
-    background-color: #333;
-    color: #fff;
-}
-
-.register-btn {
-    background-color: #FFD700;
-    color: #333;
-    margin-left: 1rem;
-}
-
-@media (max-width: 768px) {
-    .menu-toggle {
-        display: block;
-    }
-
-    .header nav ul {
-        display: none;
-        flex-direction: column;
-        background-color: #fff;
-        position: absolute;
-        top: 60px;
-        right: 0;
-        width: 100%;
-        padding: 1rem;
-    }
-    
-
-    .header.nav-open ul {
-        display: flex;
-    }
-    @media (max-width: 768px) {
-    .header.nav-open ul {
-        display: flex;
-        width:100%;
-        flex-direction: row;
-        justify-content: end;
-    }
-}
-
-    .header nav ul li {
-        margin: 1rem 0;
-    }
-}
 
 
 /* Interação dos botões ao passar o mouse */
